Rhianna   cookbook view

+++ CookBook/Views/CookBookView.swift
@@ -14,19 +14,23 @@ struct CookBookView: View {
     
     var body: some View {
         VStack{
-            List {
-                ForEach(recipes) { recipe in
-                NavigationLink {
-                    RecipeDetailView(recipe: recipe)
-                } label: {
-                    RecipeRowView(recipe: recipe)
-                }
-            }.onDelete(perform: { indexSet in
-                    for index in indexSet {
-                        modelContext.delete(recipes[index])
-                    }
-                })
-            }.listRowSpacing(10)
+            if (recipes.isEmpty){
+                Text("No saved Recipes")
+            } else {
+                List {
+                    ForEach(recipes) { recipe in
+                        NavigationLink {
+                            RecipeDetailView(recipe: recipe)
+                        } label: {
+                            RecipeRowView(recipe: recipe)
+                        }
+                    }.onDelete(perform: { indexSet in
+                        for index in indexSet {
+                            modelContext.delete(recipes[index])
+                        }
+                    })
+                }.listRowSpacing(10)
+            }
         }.navigationTitle("CookBook")
     }
 }
